requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
attrs==24.3.0
boto3==1.36.0
botocore==1.36.0
build==1.2.2
byu-pytest-utils==0.7.13
CacheControl==0.14.0
canvasapi==3.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
cleo==2.1.0
colorama==0.4.6
contourpy==1.3.1
crashtest==0.4.1
cycler==0.12.1
discord==2.3.2
discord.py==2.4.0
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.0
fonttools==4.55.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jiter==0.8.2
jmespath==1.0.1
keyring==24.3.1
kiwisolver==1.4.8
matplotlib==3.10.0
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
numpy==2.2.1
openai==1.59.7
packaging==24.2
pandas==2.2.3
pexpect==4.9.0
pillow==11.1.0
pkginfo==1.11.1
platformdirs==4.3.2
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
propcache==0.2.1
ptyprocess==0.7.0
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
pyproject_hooks==1.1.0
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
quest.py==0.2.2
rapidfuzz==3.9.7
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.0
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2024.7.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
virtualenv==20.26.4
yarl==1.18.3
zipp==3.20.1
